refactor(issue-237): crear PrivateEvents.Tests y reparar los paths de deploy
DetalleFranjaOrdinariaIgualdadTests y su IgualdadTestBase pasan a un
proyecto propio, con carpeta y namespace que espejan el ensamblado bajo
prueba. El .csproj referencia solo PrivateEvents -- que arrastra
PublicEvents, dependencia legitima del grafo -- para que la barrera sea no
alcanzar ningun DomainEvents ni Function App.

Con eso Contracts.Tests queda vacio y se elimina del disco y del slnx.
Cada ensamblado de eventos tiene ahora su proyecto de tests.

Repara ademas un defecto de despliegue que el refactor habia dejado en los
filtros paths de los workflows, invisible para el build y la suite:

- deploy-programacion y deploy-control-horas filtraban por
  src/Bitakora.ControlAsistencia.Contracts/**, un directorio que ya no
  existe. Un cambio en PublicEvents, PrivateEvents o un DomainEvents no
  habria disparado el deploy de la Function App que lo consume.
- deploy-projections nunca declaro los DomainEvents, y el worker compila
  contra ellos desde que se creo el grafo. Habria quedado desplegado con
  una version vieja de los tipos que deserializa.

Es la staleness silenciosa que esos mismos workflows ya documentaban para
global.json.

Suite igual al baseline: 473 tests (309 + 113 + 28 + 13 + 10).
